+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Change Log
+## 0.1.0
+
+- Functions not related to setting up the environment have been moved to Launch & Intellisense Configurator for Intel® oneAPI Toolkits;
+
+- Added setting up multiple environments for MultiWorkspace;
+
+- Added support for the SETVARS_CONFIG file to configure the behavior of setvars;
+
+- Updated user dialogs
+
 ##  0.0.51
 
 - Initial release